More improvements to OpenCL code.  Refs #20.
Added the pickle files and images to .gitignore because they distracted me.

Fixed --pickle option to default to False in PrepRegion.py and added
support for it in region.py.

Changed bathy.cl to only check north south east and west instead of all eight
neighbors.  Results look good and take less time.

Cleaned up pickling references in idt, elev, and bathy.

Also added support for generating images to display differences for
testing.  This same technique was used very early on with TopoMC when
it first started almost two years ago.  Hee.

Changed elev to use floats instead of ints, and now it matches the old
code.

Changed the tree generation to use ravel() instead of flatten() which
should speed it up.

+++ elev.cl
@@ -1,4 +1,4 @@
-__kernel void elev(__global int *retvals, __global int *values, const int lenvalues, const int trim, const int vscale, const int sealevel) {
+__kernel void elev(__global float *retvals, __global float *values, const uint lenvalues, const float trim, const float vscale, const float sealevel) {
   // retvals -- OUTPUT: int array of shape (lenvalues)
   //            containing values for map
   // values -- INPUT: int array of shape (lenvalues)
